Extract 'SquashException' tests into its own test class.
